Fix tenant_quotas integer overflow for PostgreSQL (Issue #1259)
Enterprise plan monthly_token_limit (3,000,000,000) exceeds PostgreSQL
integer max value (2,147,483,647), causing overflow errors when creating
tenants.

Changes:
- Add migration to change daily_token_limit and monthly_token_limit
  from integer to bigint for PostgreSQL
- Update schema-postgres.sql and baseline_2026_06_23-postgres.sql
  to use bigint for these columns

Fixes #1259

+++ schema/schema-postgres.sql
@@ -1073,8 +1073,8 @@ ALTER SEQUENCE teams_id_seq OWNED BY teams.id;
 CREATE TABLE tenant_quotas (
     id integer NOT NULL,
     tenant_id integer NOT NULL,
-    daily_token_limit integer DEFAULT 1000000,
-    monthly_token_limit integer DEFAULT 30000000,
+    daily_token_limit bigint DEFAULT 1000000,
+    monthly_token_limit bigint DEFAULT 30000000,
     daily_request_limit integer DEFAULT 10000,
     monthly_request_limit integer DEFAULT 300000,
     max_users integer DEFAULT 100,
